x86/cpu: add support for amd64
This change provides the cpu module with the subtle differences that
apply to the processor data structures in 64-bit mode. A few bits were
added to the boot module to complete SMP initialization.

+++ arch/x86/machine/boot.h
@@ -79,7 +79,7 @@ extern unsigned long boot_ap_stack_addr;
 /*
  * Size of the trampoline code used for APs.
  */
-extern unsigned long boot_mp_trampoline_size;
+extern uint32_t boot_mp_trampoline_size;
 
 /*
  * Address of the MP trampoline code.
